requirements.txt
gunicorn
eventlet
psycopg2-binary
djangorestframework
numpy
django
